Moreno45/feature/kosh task
