--- CZ/ball_bounce/03_simulation_ensembles/ball_bounce_suite_merlin_simulation_ensembles.yaml
@@ -1,31 +1,0 @@
-description:
-    name: ball-bounce 
-    description: A workflow that simulates a ball bouncing in a box over several input sets.
-
-env:
-    variables:
-        OUTPUT_PATH: ./03_simulation_ensembles/data
-        SIM_SCRIPT_PATH: ../ball_bounce.py
-
-user:
-    study:
-        run:
-            run_ball_bounce: &run_ball_bounce
-                cmd: |
-                  python $(SPECROOT)/$(SIM_SCRIPT_PATH) --output output.dsv --xpos $(X_POS_INITIAL) --ypos $(Y_POS_INITIAL) --zpos $(Z_POS_INITIAL) --xvel $(X_VEL_INITIAL) --yvel $(Y_VEL_INITIAL) --zvel $(Z_VEL_INITIAL) --gravity $(GRAVITY) --box_side_length $(BOX_SIDE_LENGTH) --group $(GROUP_ID) --run $(RUN_ID)
-                max_retries: 1
-
-study:
-    - name: run-ball-bounce
-      description: Run a family of simulations of a ball in a box. 
-      run:
-          <<: *run_ball_bounce
-
-merlin:
-    resources:
-        task_server: celery
-        overlap: False
-        workers:
-            ball_bounce_worker:
-                args: -l INFO --concurrency 4 --prefetch-multiplier 2 -O fair
-                steps: [run-ball-bounce]
