--- CZ/ball_bounce/requirements.txt
@@ -1,12 +1,0 @@
-jupyter
-llnl-sina
-kosh
-matplotlib
-maestrowf
-merlin
-seaborn
-nbconvert
-scipy
-h5py
-llnl-ibis
-trata
