+++ CZ/ball_bounce_vvuq/05_post-process_data/ball_bounce_suite_maestro_post-process_data.yaml
@@ -1,0 +1,34 @@
+description:
+    name: ball-bounce 
+    description: A workflow that simulates a ball bouncing in a box over several input sets.
+
+env:
+    variables:
+        OUTPUT_PATH: ./03_simulation_ensembles/data
+        SIM_SCRIPT_PATH: ../ball_bounce.py
+        PROCESS_SCRIPT_PATH: ../dsv_to_sina.py
+        OUTPUT_DS_PATH: ../04_manage_data/data/ensembles_output.sqlite
+        POST_PROCESS_SCRIPT_PATH: ./visualization_ensembles_sina.py
+
+study:
+    - name: run-ball-bounce
+      description: Run a family of simulations of a ball in a box. 
+      run:
+          cmd: |
+            mkdir $(SPECROOT)/../03_simulation_ensembles/data
+            python $(SPECROOT)/$(SIM_SCRIPT_PATH) --output output.dsv --xpos $(X_POS_INITIAL) --ypos $(Y_POS_INITIAL) --zpos $(Z_POS_INITIAL) --xvel $(X_VEL_INITIAL) --yvel $(Y_VEL_INITIAL) --zvel $(Z_VEL_INITIAL) --gravity $(GRAVITY) --box_side_length $(BOX_SIDE_LENGTH) --group $(GROUP_ID) --run $(RUN_ID)
+    
+    - name: ingest-ball-bounce
+      description: Ingest the outputs from the previous step
+      run:
+          cmd: |
+            mkdir $(SPECROOT)/../04_manage_data/data
+            python $(SPECROOT)/$(PROCESS_SCRIPT_PATH) $(OUTPUT_PATH) $(SPECROOT)/$(OUTPUT_DS_PATH)
+          depends: [run-ball-bounce_*]
+      
+    - name: post_process-ball-bounce
+      description: Post-process the simulation data
+      run:
+          cmd: |
+            python $(SPECROOT)/$(POST_PROCESS_SCRIPT_PATH) $(SPECROOT)
+          depends: [ingest-ball-bounce]
